+++ CHANGELOG.md
@@ -69,6 +69,7 @@
 - Re-wording of Bows/Crossbows descriptin in compendium
 - Fix Item Sheet upgrade display count to correctly show upgrade usage
 - Greeter text (the post-migration pop-up window) is displayed as HTML rather than plain text
+- Changed migration status to use a status bar instead of flooding the screen with status messages
 
 
 **Bug Fixes**